requirements.txt
asgiref==3.3.1
authentication==1.1.0
certifi==2020.11.8
click==7.1.2
cloudinary==1.22.0
dj-database-url==0.5.0
Django==3.1.3
django-tinymce==3.1.0
djangorestframework==3.12.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
passlib==1.7.4
psycopg2==2.8.6
pycryptodome==3.9.9
pytz==2020.4
redis==3.5.3
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.1
Werkzeug==1.0.1
WTForms==2.3.3
